% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\name{competition}
\alias{competition}
\title{Simulation of competition between two plant species}
\usage{
competition(row, col, time.step, info, plants)
}
\arguments{
\item{row}{Row location}

\item{col}{Column location}

\item{time.step}{Numeric value indicating the number of times steps that should be looped over (Default: 5).
Maximum number of iterations is 1000.}

\item{info}{A list including reproduction, survival, and competition probabilities, as well as, species names.
This list is generated using the \code{setup.plants} function.}

\item{plants}{An array of plant matrice with time as the depth dimension.
Array generated using the \code{run.plant.ecosystem}.}
}
\value{
an array called "plants"
}
\description{
This function take the probabilities included in our \code{info} list and applies them to the elements in the array to determine which species would win a fight between two possible species.
}
\author{
Mallory Hagadorn
}

