% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{diamond.step}
\alias{diamond.step}
\title{Apply diamond step algorithm to a general matrix}
\usage{
diamond.step(g.mat = g.mat)
}
\arguments{
\item{g.mat}{General matrix generate using the \code{general.matrix} function (Default: g.mat).}
}
\value{
a general matrix (g.mat)
}
\description{
Adding values to a general matrix (g.mat) in a diamond algorithm fashion
}

