% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pack_Test.R
\name{double.numb}
\alias{double.numb}
\title{Function for doubling a number "x"}
\usage{
double.numb(x)
}
\arguments{
\item{x}{Numeric vector (minimum length one) that is the input into \code{double.numb}.}
}
\value{
a numeric vector of length that is equivalent to the length of argument x
}
\description{
Function for doubling a number "x"
}
\examples{
x <- seq(1, 10, by=2)
x.doubled <- double.numb(x)
}

