% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\name{fight}
\alias{fight}
\title{Simulation of competition between two plant species}
\usage{
fight(name, info, plants)
}
\arguments{
\item{name}{Name of the elements that will be interacting}

\item{info}{A list including reproduction, survival, and competition probabilities, as well as, species names.
This list is generated using the \code{setup.plants} function.}

\item{plants}{An array of plant matrice with time as the depth dimension.
Array generated using the \code{run.plant.ecosystem}.}
}
\value{
an array called "plants"
}
\description{
This function take the probabilities included in our \code{info} list and applies them to the elements in the array to determine which species would win a fight between two possible species.
}
\author{
Mallory Hagadorn
}

