% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{general.matrix}
\alias{general.matrix}
\title{Generation of a terrain matrix that will be fed into plant simulation
Make starting terrain matrix with the corners seeded}
\usage{
general.matrix(n = 5, x.bar = 500)
}
\arguments{
\item{n}{Size of the grid will be a (2^n)+1 (by default: n=5; a 33 by 33 grid)}

\item{x.bar}{Numeric vector (Default: 500) that will be input as the mean into the rnorm() function.
Here, x.bar will be used to add random noise to our terrain matrix through standard deviations from the x.bar value.
This is applied to rnorm() as sd=(x.bar*x.bar), so one half the x.bar will be the standard deviation used.}
}
\value{
general matrix called g.mat
}
\description{
Function to make the initial terrain matrix with seeded corners. Ultimately, this matrix will be input into both the \code{diamond.step} and \code{square.step} functions.
}
\author{
Mallory Hagadorn
}

