% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\name{mk.plant.array}
\alias{mk.plant.array}
\title{Make empty array}
\usage{
mk.plant.array(terrain, num.timesteps = 5)
}
\arguments{
\item{terrain}{A matrix containing numeric elements that are indicates as heights and if lake.na=TRUE \code{NA}s indicated cells that are waterlogged. Terrain matrix is visualized using \code{image}.}

\item{num.timesteps}{Numeric value indicating the number of times steps that should be iterated.}
}
\value{
an array called "plants"
}
\description{
Preallocate an array which corresponds to the dimensions of terrain and the number of times steps to iterate through.
}
\author{
Mallory Hagadorn
}

