% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\name{plant.timestep}
\alias{plant.timestep}
\title{Simulate plant ecosystem for distinct points in time}
\usage{
plant.timestep(plants = plants, info = info)
}
\arguments{
\item{plants}{An array of plant matrice with time as the depth dimension.
Array generated using the \code{run.plant.ecosystem}.}

\item{info}{A list including reproduction, survival, and competition probabilities, as well as, species names.
This list is generated using the \code{setup.plants} function.}
}
\value{
an array called "plants"
}
\description{
Function that incorperates intervals of time and for our entire plant matrix
}

