% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\name{run.plant.ecosystem}
\alias{run.plant.ecosystem}
\title{Simulation of plant ecosystem through time}
\usage{
run.plant.ecosystem(terrain = terrain, num.timesteps = 5, info = info)
}
\arguments{
\item{terrain}{Numeric matrix generated using \code{make.terrain}}

\item{num.timesteps}{Numeric value indicating the number of times steps that should be looped over (Default: 5).
Maximum number of iterations is 1000.}

\item{info}{A list including reproduction, survival, and competition probabilities, as well as, species names.
This list is generated using the \code{setup.plants} function.}
}
\value{
a plant array
}
\description{
A wrapper function that generates, seeds, and places NA's in the array where they correspond in the terrain matrix.  Applies survice, reproduction, and competition functions.
}
\author{
Mallory Hagadorn
}

