% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\name{setup.plants}
\alias{setup.plants}
\title{Step up a list of plant information}
\usage{
setup.plants(repro = c(0.5, 0.5), survive = c(0.5, 0.5), comp.mat,
  name = c("a", "b"))
}
\arguments{
\item{repro}{Numeric vector of length two representing probabilities of reproduction for the corresponding species (Default: 0.5).
Reproduction probabilities should be between zero and one.
Zero represents a probability of reproduction equal to zero, or no chance of reproducing.
One represents a probability of reproduction equal to one, or 100% chance of reproduction.
In relation to \code{name}, repro vector positions 1 and 2 correspond to species "a" and "b", respectively.}

\item{survive}{Numeric vector of length two representing probabilites of survival for the corresponding species (Default: 0.5).
See variable repro for a description of probability ranges.
In relation to \code{name}, survive vector positions 1 and 2 correspond to species "a" and "b", respectively.}

\item{comp.mat}{2 by 2 matrix that corresponds to individual probabilites of a species winning the competition between another species
There is no default for this input; therefore, the user must define this variable.}

\item{name}{Character vector of length two that incorperates the two plant species names to be used in the simulation (Default: "a", "b")}
}
\value{
list of elements corresponding to argument inputs
}
\description{
This function generates a list of properties associated with each plant species used in the simulation.
}

