% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{square.step}
\alias{square.step}
\title{Apply square step algorithm to a general matrix}
\usage{
square.step(g.mat = g.mat, x.bar = 500)
}
\arguments{
\item{g.mat}{General matrix generate using the \code{general.matrix} function (Default: g.mat).}

\item{x.bar}{Numeric vector (Default: 500) that will be input as the mean into the rnorm() function.
Here, x.bar will be used to add random noise to our terrain matrix through standard deviations from the x.bar value.
This is applied to rnorm() as sd=(x.bar/2), so one half the x.bar will be the standard deviation used.}
}
\value{
a general matrix (g.mat)
}
\description{
Adding values to a general matrix (g.mat) in a square algorithm fashion
}

