% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\name{survive.fun}
\alias{survive.fun}
\title{Determines if the input (where appropriate) will survive}
\usage{
survive.fun(cell, info)
}
\arguments{
\item{cell}{A specific cell [row, column] in a matrix.
The survive function will be included in a \code{plant.timestep} function, which is applied to an array.}

\item{info}{A list including reproduction, survival, and competition probabilities, as well as, species names.
This list is generated using the \code{setup.plants} function.}
}
\value{
single element of a matrix
}
\description{
This function takes the specific cell of an input matrix (can be NA, '', or name) and determines if that cell will "survive" to the next time step.
Inputs of NA represent a waterlogged cell, whereas those indicated with '', represent cells that are unoccupied.
When a cell is filled with a specific name, that name is referenced back to the list created in \code{setup.plants}. This list contains survival probabilites, which are compared to a random number generated from a uniform distribution.
If what is in the cell survives, it will be passed to subsequent time series.
}
\author{
Mallory Hagadorn
}

