% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{make.terrain}
\alias{make.terrain}
\title{Make a terrain matrix of numeric values using the Diamond-Square Step Algorithm}
\usage{
make.terrain(n = 5, x.bar = 500, water = TRUE)
}
\arguments{
\item{n}{Size of the grid will be a (2^n)+1 (by default: n=5; a 33 by 33 grid)}

\item{x.bar}{Numeric vector (Default: 500) that will be input as the mean into the rnorm() function.
Standard deviations used in the rnorm() function are generated from the mean value.
The first standard deviation value (calculated as mean * mean) will be used in \code{general.matrix} to generate  the random corner values from a normal distribution.
The second standard deviation value (calculated as mean/2) will be used in \code{square.step} for generating noise at the finer spatial scales.}

\item{water}{Logical (Default: TRUE) that specifies whether to make all terrain values lower than zero height underwater.}
}
\value{
a terrain matrix; numeric elements that are indicates as heights and if lake.na=TRUE \code{NA}s indicated cells that are waterlogged. Terrain matrix is visualized using \code{image}.
}
\description{
A light wrapper function around \code{diamond.square.step} and the addition of lakes to the terrain matrix
}
\examples{
terrain <- make.terrain(3, 100, water = TRUE)
   #Returned is a matrix (and image) using 3 as the size indicator, 100 as the mean, and saying that any value less than or equal to 0 should be made an NA.
y <- make.terrain(water = FALSE)
   #Returned is a matrix (and image) which using the default n and x.bar values, but that specifies values less than or equal to 0 shouldn't be replaced with NA's
   #This gives the user some flexibility to not force them to use NA's
z <- make.terrain(2, 10)
   #Returned is a matrix (and image) where the n(2) and x.bar(10) are much smaller, but NA's still replace the values less than or equal to 0
}
\author{
Mallory Hagadorn
}

